add ranking
frontendアプリケーションにrecommend機能を追加
とりあえずformから酒の種類も受け取ってます
recommendアプリケーション作成も試したが、urlsが通らなかったためfrontendにまとめた
