+++ workspace/project/frontend/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from frontend.models import Sample,Quiz
 
 # Register your models here.
+
+admin.site.register(Sample)
+admin.site.register(Quiz)
