+++ workspace/project/frontend/models.py
@@ -1,3 +1,105 @@
 from django.db import models
 
+class Quiz(models.Model):
+    kindSake = models.IntegerField()
+    firstfeeling = models.IntegerField()
+    secondfeeling = models.IntegerField()
+    sex = models.IntegerField()
+
 # Create your models here.
+class Sample(models.Model):
+    # sampleid : INTEGER型で、主キー
+    sampleid = models.IntegerField(primary_key=True)
+    # sex : 0が男　1が女
+    sex = models.IntegerField(default=0)
+    # age : 年代　20代:0　30代:1　40代:2　50以上:3
+    age = models.IntegerField(default=0)
+
+    #=================以下酒パラメータ=================
+    beer = models.IntegerField(default=1)
+    weakTyuhi = models.IntegerField(default=1)  
+    mediumYyuhi = models.IntegerField(default=1)
+    strongTyuhi = models.IntegerField(default=1)
+    wisky = models.IntegerField(default=1)
+    spirits = models.IntegerField(default=1)
+    wine = models.IntegerField(default=1)
+    fruitSake = models.IntegerField(default=1)
+    liqueur = models.IntegerField(default=1)
+    umesyu = models.IntegerField(default=1)
+    sake = models.IntegerField(default=1)
+    shoTyu = models.IntegerField(default=1)
+    #=================以下野菜・肉・魚介=================
+    saladkei = models.IntegerField(default=1)
+    yasaisutelikku = models.IntegerField(default=1)
+    otukemono = models.IntegerField(default=1)
+    kimuchi = models.IntegerField(default=1)
+    rakkyou = models.IntegerField(default=1)
+    ume = models.IntegerField(default=1)
+    edamame = models.IntegerField(default=1)
+    saladchiken = models.IntegerField(default=1)
+    winnner = models.IntegerField(default=1)
+    bacon = models.IntegerField(default=1)
+    smoketang = models.IntegerField(default=1)
+    namahamu = models.IntegerField(default=1)
+    karaage = models.IntegerField(default=1)
+    tukune = models.IntegerField(default=1)
+    kamaboko = models.IntegerField(default=1)
+    kanikama = models.IntegerField(default=1)
+    tikuwa = models.IntegerField(default=1)
+    fish = models.IntegerField(default=1)
+    takobutu = models.IntegerField(default=1)
+    #=================以下惣菜・麺類・ホットスナック=================
+    gohanmono = models.IntegerField(default=1)
+    hanbagu = models.IntegerField(default=1)
+    gyouza = models.IntegerField(default=1)
+    syuumai = models.IntegerField(default=1)
+    subuta = models.IntegerField(default=1)
+    tikuzenni = models.IntegerField(default=1)
+    oden = models.IntegerField(default=1)
+    kandume = models.IntegerField(default=1)
+    udon = models.IntegerField(default=1)
+    soumen = models.IntegerField(default=1)
+    soba = models.IntegerField(default=1)
+    tyukamen = models.IntegerField(default=1)
+    kappumen = models.IntegerField(default=1)
+    karaagekun = models.IntegerField(default=1)
+    kokrokke = models.IntegerField(default=1)
+    tyuukaman = models.IntegerField(default=1)
+    amerikandoggu = models.IntegerField(default=1)
+    yakutori = models.IntegerField(default=1)
+
+    #=================以下洋・和菓子=================
+    purin = models.IntegerField(default=1)
+    taruto = models.IntegerField(default=1)
+    warbimoti = models.IntegerField(default=1)
+    mannjyuu = models.IntegerField(default=1)
+    dorayaki = models.IntegerField(default=1)
+    kukki = models.IntegerField(default=1)
+    rolecake = models.IntegerField(default=1)
+    baumukuuhen = models.IntegerField(default=1)
+    dounatu = models.IntegerField(default=1)
+    masyumaro = models.IntegerField(default=1)
+    chocolategasi = models.IntegerField(default=1)
+    poppukon = models.IntegerField(default=1)
+    soltsnack = models.IntegerField(default=1)
+    spicysnack = models.IntegerField(default=1)
+    konsomesnack = models.IntegerField(default=1)
+    agegasi = models.IntegerField(default=1)
+    tiitara = models.IntegerField(default=1)
+    mamegasi = models.IntegerField(default=1)
+    osenbay = models.IntegerField(default=1)
+    surume = models.IntegerField(default=1)
+    beefjurky = models.IntegerField(default=1)
+
+    #=================乳製品・フルーツ・アイス=================
+    onsentamago = models.IntegerField(default=1)
+    sakeruchees = models.IntegerField(default=1)
+    smokecheez = models.IntegerField(default=1)
+    tofu = models.IntegerField(default=1)
+    dryfruits = models.IntegerField(default=1)
+    fruits = models.IntegerField(default=1)
+    vanilaice = models.IntegerField(default=1)
+    chocoice = models.IntegerField(default=1)
+    syabettoice = models.IntegerField(default=1)
+    fruitice = models.IntegerField(default=1)
+
