+++ workspace/project/frontend/templates/frontend/new.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+ 
+<head>
+    <meta charset="utf-8" />
+    <title>フォーム</title>
+</head>
+ 
+<body>
+    <h1>クイズ！！！</h1>
+    <form method="POST" novalidate>
+        {% csrf_token %}
+        {{ form.as_p }}
+        <button type="submit">送信</button>
+    </form>
+    <p>{{ message }}</p>
+</body>
+ 
+</html>
