+++ workspace/project/frontend/urls.py
@@ -7,6 +7,8 @@ urlpatterns = [
     path('submit', views.submit, name="submit"),
     path('quiz', views.quiz, name="quiz"),
     path('photo', views.photo, name="photo"),
-    path('result', views.result, name="result")
+    path('result', views.result, name="result"),
+    path('recommend', views.recommend, name="recommend"),
+    path('new', views.new, name="new"),
 
 ]
