+++ workspace/project/project/settings.py
@@ -38,7 +38,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'frontend',
-    'recommend',
 ]
 
 MIDDLEWARE = [
