+++ workspace/project/project/urls.py
@@ -22,7 +22,9 @@ urlpatterns = [
     path('quiz', include('frontend.urls')),
     path('photo', include('frontend.urls')),
     path('result', include('frontend.urls')),
-    path('recommend', include('recommend.urls')),
-    path('admin', admin.site.urls),
+    path('recommend', include('frontend.urls')),
+    path('new', include('frontend.urls')),
 
+
+    path('admin/', admin.site.urls),
 ]
