--- workspace/project/recommend/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from . import views
-
-app_name="recommend"
-
-urlpatterns = [
-    path('recommend', views.recommend, name="recommend"),
-]
